+++ src/scrapy_redis/spiders.py
@@ -1,11 +1,13 @@
-from scrapy import signals
+import json
+from collections.abc import Iterable
+from scrapy import signals, FormRequest
 from scrapy.exceptions import DontCloseSpider
 from scrapy.spiders import Spider, CrawlSpider
-from collections import Iterable
+from scrapy_redis.utils import TextColor
 import time
 
 from . import connection, defaults
-from .utils import bytes_to_str
+from .utils import bytes_to_str, is_dict
 
 
 class RedisMixin(object):
@@ -19,6 +21,7 @@ class RedisMixin(object):
 
     # Idle start time
     spider_idle_start_time = int(time.time())
+    max_idle_time = None
 
     def start_requests(self):
         """Returns a batch of start requests from redis."""
@@ -54,11 +57,7 @@ class RedisMixin(object):
             raise ValueError("redis_key must not be empty")
 
         if self.redis_batch_size is None:
-            # TODO: Deprecate this setting (REDIS_START_URLS_BATCH_SIZE).
-            self.redis_batch_size = settings.getint(
-                'REDIS_START_URLS_BATCH_SIZE',
-                settings.getint('CONCURRENT_REQUESTS'),
-            )
+            self.redis_batch_size = settings.getint('CONCURRENT_REQUESTS', defaults.REDIS_CONCURRENT_REQUESTS)
 
         try:
             self.redis_batch_size = int(self.redis_batch_size)
@@ -69,20 +68,31 @@ class RedisMixin(object):
             self.redis_encoding = settings.get('REDIS_ENCODING', defaults.REDIS_ENCODING)
 
         self.logger.info("Reading start URLs from redis key '%(redis_key)s' "
-                         "(batch size: %(redis_batch_size)s, encoding: %(redis_encoding)s",
+                         "(batch size: %(redis_batch_size)s, encoding: %(redis_encoding)s)",
                          self.__dict__)
 
         self.server = connection.from_settings(crawler.settings)
 
-        if self.settings.getbool('REDIS_START_URLS_AS_SET', defaults.START_URLS_AS_SET):
+        if settings.getbool('REDIS_START_URLS_AS_SET', defaults.START_URLS_AS_SET):
             self.fetch_data = self.server.spop
             self.count_size = self.server.scard
-        elif self.settings.getbool('REDIS_START_URLS_AS_ZSET', defaults.START_URLS_AS_ZSET):
+        elif settings.getbool('REDIS_START_URLS_AS_ZSET', defaults.START_URLS_AS_ZSET):
             self.fetch_data = self.pop_priority_queue
             self.count_size = self.server.zcard
         else:
             self.fetch_data = self.pop_list_queue
             self.count_size = self.server.llen
+
+        if self.max_idle_time is None:
+            self.max_idle_time = settings.get(
+                "MAX_IDLE_TIME_BEFORE_CLOSE",
+                defaults.MAX_IDLE_TIME
+            )
+
+        try:
+            self.max_idle_time = int(self.max_idle_time)
+        except (TypeError, ValueError):
+            raise ValueError("max_idle_time must be an integer")
 
         # The idle signal is called when the spider has no requests left,
         # that's when we will schedule new requests from redis queue
@@ -131,16 +141,37 @@ class RedisMixin(object):
                 yield reqs
                 found += 1
             else:
-                self.logger.debug("Request not made from data: %r", data)
+                self.logger.debug(f"Request not made from data: {data}")
 
         if found:
-            self.logger.debug("Read %s requests from '%s'", found, self.redis_key)
+            self.logger.debug(f"Read {found} requests from '{self.redis_key}'")
 
     def make_request_from_data(self, data):
-        """Returns a Request instance from data coming from Redis.
+        """
+        Returns a `Request` instance for data coming from Redis.
 
-        By default, ``data`` is an encoded URL. You can override this method to
-        provide your own message decoding.
+        Overriding this function to support the `json` requested `data` that contains
+        `url` ,`meta` and other optional parameters. `meta` is a nested json which contains sub-data.
+
+        Along with:
+        After accessing the data, sending the FormRequest with `url`, `meta` and addition `formdata`, `method`
+        For example:
+        {
+            "url": "https://exaple.com",
+            "meta": {
+                'job-id':'123xsd',
+                'start-date':'dd/mm/yy'
+            },
+            "url_cookie_key":"fertxsas",
+            "method":"POST"
+        }
+
+        If `url` is empty, return []. So you should verify the `url` in the data.
+        If `method` is empty, the request object will set method to 'GET', optional.
+        If `meta` is empty, the request object will set `meta` to {}, optional.
+
+        This json supported data can be accessed from 'scrapy.spider' through response.
+        'request.url', 'request.meta', 'request.cookies', 'request.method'
 
         Parameters
         ----------
@@ -148,8 +179,24 @@ class RedisMixin(object):
             Message from redis.
 
         """
-        url = bytes_to_str(data, self.redis_encoding)
-        return self.make_requests_from_url(url)
+        formatted_data = bytes_to_str(data, self.redis_encoding)
+
+        if is_dict(formatted_data):
+            parameter = json.loads(formatted_data)
+        else:
+            self.logger.warning(f"{TextColor.WARNING}WARNING: String request is deprecated, please use JSON data format. \
+                Detail information, please check https://github.com/rmax/scrapy-redis#features{TextColor.ENDC}")
+            return FormRequest(formatted_data, dont_filter=True)
+
+        if parameter.get('url', None) is None:
+            self.logger.warning(f"{TextColor.WARNING}The data from Redis has no url key in push data{TextColor.ENDC}")
+            return []
+
+        url = parameter.pop("url")
+        method = parameter.pop("method").upper() if "method" in parameter else "GET"
+        metadata = parameter.pop("meta") if "meta" in parameter else {}
+
+        return FormRequest(url, dont_filter=True, method=method, formdata=parameter, meta=metadata)
 
     def schedule_next_requests(self):
         """Schedules a request if available"""
@@ -168,9 +215,8 @@ class RedisMixin(object):
 
         self.schedule_next_requests()
 
-        max_idle_time = self.settings.getint("MAX_IDLE_TIME_BEFORE_CLOSE")
         idle_time = int(time.time()) - self.spider_idle_start_time
-        if max_idle_time != 0 and idle_time >= max_idle_time:
+        if self.max_idle_time != 0 and idle_time >= self.max_idle_time:
             return
         raise DontCloseSpider
 
@@ -202,8 +248,8 @@ class RedisSpider(RedisMixin, Spider):
     """
 
     @classmethod
-    def from_crawler(self, crawler, *args, **kwargs):
-        obj = super(RedisSpider, self).from_crawler(crawler, *args, **kwargs)
+    def from_crawler(cls, crawler, *args, **kwargs):
+        obj = super(RedisSpider, cls).from_crawler(crawler, *args, **kwargs)
         obj.setup_redis(crawler)
         return obj
 
@@ -234,7 +280,7 @@ class RedisCrawlSpider(RedisMixin, CrawlSpider):
     """
 
     @classmethod
-    def from_crawler(self, crawler, *args, **kwargs):
-        obj = super(RedisCrawlSpider, self).from_crawler(crawler, *args, **kwargs)
+    def from_crawler(cls, crawler, *args, **kwargs):
+        obj = super(RedisCrawlSpider, cls).from_crawler(crawler, *args, **kwargs)
         obj.setup_redis(crawler)
         return obj